+++ packages/frontier-plugins/plugin-vue/src/commands/vue/add/component.ts
@@ -1,11 +1,11 @@
 import { Command, flags } from '@oclif/command';
 import chalk from 'chalk';
-import { toKebabCase } from '@rdfrontier/stdlib';
+import { toKebabCase } from '@jantaeleckie/stdlib';
 import { checkProjectValidity, parseComponentName } from '../../../utils/utilities';
 import { CLI_COMMANDS, CLI_STATE, DOCUMENTATION_LINKS } from '../../../utils/constants';
 import { addElementFunction } from '../../../functions/addElement';
-import { invalidProject } from '@rdfrontier/plugin-shared'
-import { catchError } from '@rdfrontier/plugin-shared';
+import { invalidProject } from '@jantaeleckie/plugin-shared'
+import { catchError } from '@jantaeleckie/plugin-shared';
 
 const TEMPLATE_FOLDERS = ['component'];