+++ packages/frontier-core/cli/CHANGELOG.md
@@ -1,8 +1,8 @@
-# @rdfrontier/cli
+# @jantaeleckie/cli
 
 ## 0.0.3
 
 ### Patch Changes
 
 - Updated dependencies [286d4e6]
-  - @rdfrontier/stdlib@0.0.2
+  - @jantaeleckie/stdlib@0.0.2