+++ packages/frontier-plugins/plugin-shared/CHANGELOG.md
@@ -1,8 +1,8 @@
-# @rdfrontier/plugin-shared
+# @jantaeleckie/plugin-shared
 
 ## 0.0.2
 
 ### Patch Changes
 
 - Updated dependencies [286d4e6]
-  - @rdfrontier/stdlib@0.0.2
+  - @jantaeleckie/stdlib@0.0.2