+++ packages/frontier-plugins/plugin-vue/CHANGELOG.md
@@ -1,9 +1,9 @@
-# @rdfrontier/plugin-vue
+# @jantaeleckie/plugin-vue
 
 ## 1.0.1
 
 ### Patch Changes
 
 - Updated dependencies [286d4e6]
-  - @rdfrontier/stdlib@0.0.2
-  - @rdfrontier/plugin-shared@0.0.2
+  - @jantaeleckie/stdlib@0.0.2
+  - @jantaeleckie/plugin-shared@0.0.2